+++ end_to_end_tests/golden-master/my_test_api_client/api/default.py
@@ -5,20 +5,16 @@ import httpx
 
 from ..client import AuthenticatedClient, Client
 from ..errors import ApiResponseError
-from ..models.abc_response import ABCResponse
 
 
-def ping_ping_get(
-    *, client: Client,
-) -> Union[
-    ABCResponse,
-]:
+def ping_ping_get(*, client: Client,) -> bool:
+
     """ A quick check to see if the system is running  """
     url = "{}/ping".format(client.base_url)
 
     response = httpx.get(url=url, headers=client.get_headers(),)
 
     if response.status_code == 200:
-        return ABCResponse.from_dict(cast(Dict[str, Any], response.json()))
+        return bool(response.text)
     else:
         raise ApiResponseError(response=response)