+++ end_to_end_tests/golden-master/my_test_api_client/api/users.py
@@ -17,6 +17,7 @@ def get_user_list(
 ) -> Union[
     List[AModel], HTTPValidationError,
 ]:
+
     """ Get a list of things  """
     url = "{}/tests/".format(client.base_url)
 
@@ -52,6 +53,7 @@ def upload_file_tests_upload_post(
 ) -> Union[
     None, HTTPValidationError,
 ]:
+
     """ Upload a file  """
     url = "{}/tests/upload".format(client.base_url)