+++ end_to_end_tests/golden-master/my_test_api_client/models/__init__.py
@@ -1,7 +1,6 @@
 """ Contains all the data models used in inputs/outputs """
 
 from .a_model import AModel
-from .abc_response import ABCResponse
 from .an_enum_value import AnEnumValue
 from .an_enum_value_item import AnEnumValueItem
 from .body_upload_file_tests_upload_post import BodyUploadFileTestsUploadPost